add v as an option for sample entropy

+++ plot_module/auc_as_function_of_layers.py
@@ -198,7 +198,7 @@ for i in sorted(p_values_dict.keys(),key=lambda x:x):
 out_str='*'
 text=[]
 for k in sorted(p_values_dict.keys(),key=lambda x:x):
-    text.append(k+' - $p_{value}$<5e-%d'%(np.log10(5/k)))
+    text.append(k+' - $p_{value}$<5e-%d'%(np.log10(5/k)).astype(int))
 props = dict(boxstyle='round', facecolor='wheat', alpha=0.5)
 text = '\n'.join(text)
 # place a text box in upper left in axes coords
